Update and rename jumpcutter.py to JumpCutterUltra.py
Did all kinds of changes to the original jumpcutter script.  I put in 2 methods for the script to find the FPS of the input video.  The first method was the one originally attempted in the script, looking for the fps in the para text file.  The path was wrong and the input file was set to a hardcoded name "input.mp4", so I renamed it to the variable.  Also the re.search options I fixed so that it now finds the fps.  However in case this ever becomes an issue with ffmpeg in the future, I also made it fall back to using OpenCV-Python to directly read the input file and encode at the framerate found there.  Also I changed the default audio sampling to 48000, and also created a "--extras" option that now allows the user to pass encoding options directly to ffmpeg, for example, higher bitrates for audio and video, and different types of encoding methods, CPU or GPU, for example. I call my edit "Jump Cutter Ultra"! This took me quite a bit of time but you guys rocked on this script so it was SO worth it! THANK YOU SO MUCH FOR THIS!!!!!!!!!